[MOD] Tests structure
